Restructure and switch to Maven/Tycho
The natural development environment for Eclipse Plug-ins is PDE. As nice
as dogfooding is, the investment to try Bazel reproducing the proper
output for Eclipse (p2 repository, bundles, features, etc.) is just not
worth it - especially given the loss of confort at development time.

In order to make Tycho less invasive, the pom-less option was used. This
requires a structured layout, which triggered moving source code around.

Docs update to follow.