Tycho and Restructuring (#214)
* Fix license headers and update to EPL 2.0

The Apache License headers in the files was mis-leading. The majority of
the file was EPL and cannot be re-licensed to Apache.

M2Eclipse updated to EPL v2, which we should honor.

* Create a project in the root to simplify working with BEF

* Restructure and switch to Maven/Tycho

The natural development environment for Eclipse Plug-ins is PDE. As nice
as dogfooding is, the investment to try Bazel reproducing the proper
output for Eclipse (p2 repository, bundles, features, etc.) is just not
worth it - especially given the loss of confort at development time.

In order to make Tycho less invasive, the pom-less option was used. This
requires a structured layout, which triggered moving source code around.

Fixes #6 

* Update the docs

* CI build

* Remove things no longer needed

* Add links to doc

* Put project specific settings into SCM