Always launch with Bash Environment (#441)
This fixed potential issues on MacOS when Eclipse is started as desktop app. Desktop apps have a different environment than shells.

Introduce a ShellEnvironment which allows to provide additional settings
to ShellCommandBuilder and ShellCommand. In the future this can be used
to provide more environment variables.

A preference option can be set to disable to behavior.