Use NestedSet for building file tree in BEP parsing
There were still cases in the old logic where jars would be missing from
the runtime_classpath when multiple targets are contained in the BEP.
The flat list handling and propagation of information seems to be not
working correctly in all possible cases.

This changes drops the old implementation in favor of Bazel's
NestedSets. The NestedSets allow maintaining the FileSet graph at it is
submitted in BEP. We then only need to collect information about the top
level targets and connect the NestedSets there.