Replace dynamic project dependency with inverse classpath dependency
Instead of making each project depend on the workspace project we change
the dependency direction and create a classpath dependency from the
workspace project back to the individual projects.

This also filters any generated jars from projects because we do prefer
the project as a dependency instead of the individual jars.

As a benefit, the search for references should not properly find
references in binary code which is not part of the project view. This
should help with refactorings to discover potential breakages early.