Close gaps when aspects failed to gather library details
This is a bit more tricky. I noticed ijars on the classpath when I
didn't expect them to be there. This is because we try to eliminate them
as much as possible. They don't work in IDEs nicely. Turns out that with
a little bit of code we are able to turn missing libraries into more
complete objects. These objects (especially the target label) allow a
proper translation later to source project references instead of jars
where applicable.