Massive update to IJ Aspects
The IJ Aspects progressed quite a bit. A new templating system was
introduced to create aspects with different behavior depending on the
Bazel version.

Additional the design was modified so that aspects are no longer added
as a repository but become part of the workspace, i.e. they are copied
into the workspace during sync. This seems to work better with Bazel
caching (according to the IJ Aspects commit history).

When importing aspects we no longer consume them as a repository.
Instead we build them directly in the IJ repository and then copy &
extract into our zip archive. The zip archive is later extracted into a
temp location for better working with multiple Bazel workspaces.